feat: Social links collection detail
